+++ packages/infra-common/src/index.ts
@@ -6,5 +6,6 @@
  */
 export * from "./core/orbitDomainError";
 export * from "./types/orbitDomain";
+export * from "./types/governance";
 export * from "./utils/digest";
 export * from "./utils/versionIdGen";